Ensure libdbus clients can authenticate with ibus' GDBusServer
Backported from upstream 2.62.x branch.

The regression test is not included, because it doesn't seem to be 100%
reliable in stretch: it intermittently hangs while waiting for the libdbus
call to complete, especially when running under sbuild. The cause seems
to be some sort of multi-threading issue rather than a problem with
authentication, and the test usually does pass, so it doesn't invalidate
the bug fix.

Closes: #941018