debian/tests/build: Make cross-test friendly
autopkgtest is soon to get a `-a ARCHITECTURE` switch, which will
cross-test autopkgtests. This is to be detected by the presence of the
`dpkg-architecture`-style family of variables being set in the
environment.

For build tests like `glib2.0`'s `build` test, this means that we should
test "${DEB_HOST_ARCH}" and invoke the cross toolchain as necessary.

Closes: #946355

Signed-off-by: Iain Lane <laney@debian.org>