Rework 01_gettext-desktopfiles.patch to not add new public API
Downstreams should not be adding new public API to GLib. From some code
searching, this doesn’t appear to be used in more than one or two
places, so won’t be too inconvenient to drop.

The original patch should either be upstreamed (I’d be open to some form
of it, if there’s still evidence it’s useful) or dropped. If it’s
upstreamed, the new keys should be standardised.

The alternative to this was to document the added public API; its
addition was causing the new gtk-doc tests in GLib to fail.

Signed-off-by: Philip Withnall <withnall@endlessm.com>