Mark the DBUS_COOKIE_SHA1 parts of gdbus-server-auth test as flaky
This is not reliable enough to always pass on buildds, but is too
intermittent to be able to reproduce the failure in a development
environment, and DBUS_COOKIE_SHA1 is not an important enough feature
to justify failing the build for this.

As with other flaky tests, we still run this as an autopkgtest in an
attempt to get more useful information, but we ignore failure.