d/p/debian/61_glib-compile-binaries-path.patch: Remove
This patch turns out to be unnecessary, and is harmful for
cross-compiling. glib-compile-schemas is an architecture-independent
interface provided by the libglib2.0-bin package, which is
Multi-Arch: foreign. During cross-compilation we want it to be provided
by the build architecture's libglib2.0-bin.

As an implementation detail of Debian's GLib packaging, the actual binary
that implements glib-compile-schemas is bundled with the shared libraries
in libglib2.0-0, and /usr/bin/glib-compile-schemas is a symlink to the
primary architecture's copy of the actual binary - this was done to
break a circular dependency in the runtime libraries. However,
/usr/bin/glib-compile-schemas is still the correct entry-point during
cross-compilation.

Thanks: Helmut Grohne
Closes: #982213