d/patches: Resolve integer overflows, including CVE-2021-27219
These backported patches resolve an integer overflow that is known to be
attacker-triggerable in polkit (policykit-1), as well as replacing other
simple uses of g_memdup() with g_memdup2(). Overflows in most of these
places would not be attacker-triggerable, but replacing them is simpler
than assessing whether they are attacker-triggerable.

The more complicated changes from 2.66.7 have not been backported, to avoid
regressions in Debian 10. They are not believed to be attacker-triggerable.

Closes: #982778