postinst: Recreate GIO module directory if deleted
The GIO modules directory is shipped in the .deb, so it should not
disappear, meaning we shouldn't need to guard the call to
gio-querymodules. We're running it with "|| true", so the maintainer
script will not fail in any case; and if it has somehow disappeared,
it's appropriate that we get a warning.

However, because the postrm in previous versions was overzealous about
removing an empty GIO modules directory (#987913), we should recover
by recreating it. Use install -d in preference to mkdir -p to get
predictable permissions.

Similarly, the schemas directory is shipped in the .deb and was always
only deleted during purge, so we don't need to check for it: we should
be able to compile schemas unconditionally (and if the directory has
somehow disappeared, it would be correct to get a warning).

Closes: #987913